Quick fix for consensus interface
